Time: 0 ms (100%), Space: 45.5 MB (86.49%) - LeetHub

+++ 1491-average-salary-excluding-the-minimum-and-maximum-salary/1491-average-salary-excluding-the-minimum-and-maximum-salary.java
@@ -11,7 +11,7 @@ class Solution {
             totalSum += employeeSalary;
         }
 
-        return (double) (totalSum - minimum - maximum) / (salary.length - 2);
+        return (totalSum - minimum - maximum) / (salary.length - 2);
         
     }
 }
